pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "facial-keypoints-detection"
version = "1.0.0"
description = "CNN and ResNet models for facial keypoint detection on the Kaggle benchmark"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Honghuai Ke", email = "kingke0927@gmail.com"},
]
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
pytorch = ["torch>=2.0.0"]
tensorflow = ["tensorflow>=2.13.0"]
all = ["torch>=2.0.0", "tensorflow>=2.13.0"]

[project.urls]
Homepage = "https://github.com/allureking/FacialKeypointsDetection"

[tool.setuptools.packages.find]
include = ["keypoints*"]
